Skip films with already downloaded posters
Films whose posters already exist in the posters folder are skipped,
i.e. their posters are not re-downloaded.

This comes in handy if the program was interrupted before downloading
all the posters of all the films in the csv.
This way, when the programme is re-run, posters are not re-downloaded
if they were downloaded previously.